issue-29 Move Clikt based classes to a separate package
